refactor: use project references
