+++ babel.config.mjs
@@ -1,0 +1,6 @@
+export default {
+	presets: [
+		["@babel/preset-env", { targets: { node: "current" } }],
+		"@babel/preset-typescript",
+	],
+};
