+++ eslint.config.mjs
@@ -6,7 +6,7 @@ import globals from "globals";
 
 export default tseslint.config(
 	{
-		ignores: ["dist/**", "__fixtures__/dist/**"],
+		ignores: ["dist/**", "__fixtures__/dist/**", "packages/**/build"],
 	},
 	{
 		extends: [eslint.configs.recommended],
@@ -28,6 +28,7 @@ export default tseslint.config(
 				tsconfigRootDir: import.meta.dirname,
 			},
 		},
+		// include d.ts files:
 		files: ["**/*.ts"],
 	},
 );
