+++ jest.config.mjs
@@ -1,13 +1,14 @@
-import { createJsWithTsEsmPreset } from "ts-jest";
+// import { createJsWithTsEsmPreset } from "ts-jest";
 
-const presetConfig = createJsWithTsEsmPreset({
-	//...options
-});
+// const presetConfig = createJsWithTsEsmPreset({
+// 	//...options
+// });
 
 const jestConfig = {
-	...presetConfig,
+	// ...presetConfig,
 	preset: "jest-puppeteer",
 	globalSetup: "./jest-setup.mjs",
+	testPathIgnorePatterns: ["/node_modules/", "packages/.*/build"],
 };
 
 export default jestConfig;
