+++ package.json
@@ -7,8 +7,9 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "install-puppeteer-browser": "puppeteer browsers install chrome",
-    "lint": "eslint . && prettier --list-different . && tsc",
+    "lint": "eslint . && prettier --list-different .",
     "build": "webpack",
+    "clean": "tsc --build --clean && find ./packages -type d -name \"build\" -exec rm -rf {} +",
     "format": "prettier --write .",
     "prepare": "husky"
   },
@@ -21,6 +22,9 @@
   "license": "BSD-3-Clause",
   "packageManager": "pnpm@10.6.2",
   "devDependencies": {
+    "@babel/core": "^7.26.10",
+    "@babel/preset-env": "^7.26.9",
+    "@babel/preset-typescript": "^7.27.0",
     "@eslint/js": "^9.22.0",
     "@freecodecamp/curriculum-helpers": "3.10.0",
     "@types/chai": "^4.3.20",
@@ -28,6 +32,7 @@
     "@types/enzyme-adapter-react-16": "^1.0.9",
     "@types/jest": "^29.5.14",
     "@types/jquery": "^3.5.32",
+    "babel-jest": "^29.7.0",
     "chai": "^4.5.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.8",
@@ -45,7 +50,6 @@
     "prettier": "^3.5.3",
     "process": "^0.11.10",
     "puppeteer": "^24.4.0",
-    "ts-jest": "^29.2.6",
     "ts-loader": "^9.5.2",
     "typescript": "^5.8.2",
     "typescript-eslint": "^8.26.1",
