+++ packages/frame-evaluators/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../base-tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "lib": ["dom"],
+    "outDir": "./build"
+  }
+}
