+++ packages/main/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../base-tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./build"
+  }
+}
