+++ packages/workers/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../base-tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "lib": ["WebWorker"],
+    "outDir": "./build"
+  }
+}
