+++ tsconfig.json
@@ -1,17 +1,14 @@
 {
-  "include": ["src/**/*", "__tests__/**/*"],
-  "compilerOptions": {
-    "target": "ES2024",
-    "module": "ESNext",
-    "moduleResolution": "bundler",
-    "lib": ["WebWorker", "DOM"],
-    "allowJs": true,
-    "strict": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "esModuleInterop": true,
-    "resolveJsonModule": true,
-    "noEmit": true,
-    "noUncheckedIndexedAccess": true
-  }
+  "include": ["types/*.d.ts"],
+  "references": [
+    {
+      "path": "./packages/main"
+    },
+    {
+      "path": "./packages/workers"
+    },
+    {
+      "path": "./packages/frame-evaluators"
+    }
+  ]
 }
