+++ types/test-evaluator.d.ts
@@ -17,6 +17,22 @@ export type InitEvent<Data> = MessageEvent<{
 	value: Data;
 }>;
 
+// TODO: Can/should we export these from the packages?
+export interface InitTestFrameOptions {
+	code: {
+		contents?: string;
+		editableContents?: string;
+	};
+	loadEnzyme?: boolean;
+}
+
+export interface InitWorkerOptions {
+	code: {
+		contents?: string;
+		editableContents?: string;
+	};
+}
+
 export interface TestEvaluator {
 	init(opts: unknown): Promise<void>;
 	runTest(test: string): Promise<Pass | Fail>;
