+++ types/test-runner.d.ts
@@ -1,0 +1,2 @@
+export type ReadyEvent = MessageEvent<{ type: "ready" }>;
+export type ResultEvent = MessageEvent<{ type: "result"; value: unknown }>;
